fix(hot-fix): separate cart context component on 3 file to fit linter rules, fix rule unknown promocode componen
