+++ src/components/Basket/BasketProductList/BasketProductList.tsx
@@ -2,7 +2,7 @@ import { useState, useEffect } from "react";
 import Link from "../../common/link/link";
 import { BasketProductCard } from "../BasketProductCard/BasketProductCard";
 import { useNavigate } from "react-router-dom";
-import { useCart } from "../../context/CartContext";
+import { useCart } from "../../context/useCart";
 import "./BasketProductList.scss";
 
 export function BasketProductList() {
