+++ src/components/Basket/OrderSummary/OrderPromoCode/OrderPromoCode.tsx
@@ -3,11 +3,11 @@ import { Form } from "react-router-dom";
 import InputField from "../../../common/inputField/inputField";
 import Button from "../../../common/button/button";
 import { MdOutlineDiscount } from "react-icons/md";
-import { useCart } from "../../../context/CartContext";
+import { useCart } from "../../../context/useCart";
 
 import "./OrderPromoCodeStyles.scss";
 
-export default function () {
+export default function OrderPromoCode() {
   const [promoCode, setPromococde] = useState("");
 
   const { applyPromoCode } = useCart();
