+++ src/components/context/useCart.tsx
@@ -1,0 +1,11 @@
+import { useContext } from "react";
+import { CartContext } from "./CartContext";
+import type { CartContextType } from "./CartContext";
+
+export const useCart = (): CartContextType => {
+  const context = useContext(CartContext);
+  if (!context) {
+    throw new Error("useCart must be used within a CartDataProvider");
+  }
+  return context;
+};
