+++ src/layouts/MainLayout.tsx
@@ -1,7 +1,7 @@
 import { Outlet } from "react-router-dom";
 import Header from "../components/Header/Header";
 import Footer from "../components/Footer/Footer";
-import { CartDataProvider } from "../components/context/CartContext";
+import { CartDataProvider } from "../components/context/CartContextProvider";
 
 function MainLayout() {
   return (
