+++ src/pages/Login/Login.tsx
@@ -10,7 +10,7 @@ import { validateEmailFormat, validatePasswordStrength } from "../../utils/valid
 import { AuthService } from "../../services/AuthService";
 import { signIn } from "../../services/customerService/customerService";
 import { TokenService } from "../../services/TokenService";
-import { useCart } from "../../components/context/CartContext";
+import { useCart } from "../../components/context/useCart";
 import "./Login.scss";
 
 export default function LoginPage() {
